Move getDefaultIds to metadata repository
