+++ src/data/instance/InstanceD2ApiRepository.ts
@@ -1,4 +1,3 @@
-import _ from "lodash";
 import { Instance } from "../../domain/instance/entities/Instance";
 import { InstanceMessage } from "../../domain/instance/entities/Message";
 import { User } from "../../domain/instance/entities/User";
@@ -41,27 +40,6 @@ export class InstanceD2ApiRepository implements InstanceRepository {
     @cache()
     public getBaseUrl(): string {
         return this.api.baseUrl;
-    }
-
-    @cache()
-    public async getDefaultIds(filter?: string): Promise<string[]> {
-        const response = (await this.api
-            .get("/metadata", {
-                filter: "code:eq:default",
-                fields: "id",
-            })
-            .getData()) as {
-            [key: string]: { id: string }[];
-        };
-
-        const metadata = _.pickBy(response, (_value, type) => !filter || type === filter);
-
-        return _(metadata)
-            .omit(["system"])
-            .values()
-            .flatten()
-            .map(({ id }) => id)
-            .value();
     }
 
     @cache()
